(mostly) Fixed instance and parsing errors
Fixed the default instance to be a correct input for the problem. Also fixed the parsing to work as expected to give the values back from the problem instance. The parser produces an error at some point which I haven't worked out, but the outputs from it are still good (and it was based on the weighted cut implementation, which gives an error of its own). So it was good enough for me for now.